// src/enums/indicators.enum.ts
export enum Indicator {
  SURFACE_AREA = "f271",
  ALTITUDE = 'f258',
  LONGITUDE = 'f328',
  LATITUDE = 'f329',
  MEN = 'f318',
  WOMEN = 'f320',
  TOTAL_POPULATION = 'f321',
  EMPLOYED_POPULATION = 'f221',
  UNEMPLOYED_POPULATION = 'f222',
  ACTIVE_POPULATION = 'f223',
  INACTIVE_POPULATION = 'f60'
}